Fixed bug with resize tracking when minified
`player.width` and `player.height` are functions not properties. So,
they need to be called instead of simple referenced. I was getting event
actions like the following in my analytics:

`resize - function(a,c){return E(this,"width",a,c)}*function(a,c){return
E(this,"height",a,c)}`

This should resolve the problem, but I haven't tested it yet.